Implementing package orientation
