--- software_architecture_flutter/lib/app/modules/home/repositories/apiAdvisor_repository_interface.dart
@@ -1,5 +1,0 @@
-import 'package:software_architecture_flutter/app/modules/home/models/apiAdvisor_model.dart';
-
-abstract class IApiAdvisor {
-  Future<ApiAdvisorModel?> getWeather();
-}
