+++ software_architecture_flutter/modules/core_module/LICENSE
@@ -1,0 +1,1 @@
+TODO: Add your license here.
