+++ software_architecture_flutter/modules/core_module/lib/core_module.dart
@@ -1,0 +1,9 @@
+library core_module;
+
+export 'src/interfaces/client_http_interface.dart';
+export 'src/interfaces/local_storage_interface.dart';
+export 'src/models/appConfig_model.dart';
+export 'src/models/user_model.dart';
+export 'src/services/client_http_service.dart';
+export 'src/services/shared_local_storage_service.dart';
+export 'src/viewmodels/change_theme_viewmodel.dart';
