+++ software_architecture_flutter/modules/core_module/lib/src/interfaces/client_http_interface.dart
