+++ software_architecture_flutter/modules/core_module/lib/src/interfaces/local_storage_interface.dart
