+++ software_architecture_flutter/modules/core_module/lib/src/models/appConfig_model.dart
