+++ software_architecture_flutter/modules/core_module/lib/src/models/user_model.dart
