+++ software_architecture_flutter/modules/core_module/lib/src/services/client_http_service.dart
@@ -1,5 +1,5 @@
 import 'package:dio/dio.dart';
-import 'package:software_architecture_flutter/app/core/interfaces/client_http_interface.dart';
+import 'package:core_module/src/interfaces/client_http_interface.dart';
 
 class ClientHttpService implements IClientHttp {
   final Dio dio = Dio();
