+++ software_architecture_flutter/modules/core_module/lib/src/services/shared_local_storage_service.dart
@@ -1,5 +1,5 @@
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:software_architecture_flutter/app/core/interfaces/local_storage_interface.dart';
+import 'package:core_module/src/interfaces/local_storage_interface.dart';
 
 class SharedLocalStorageService implements ILocalStorage {
   @override
