+++ software_architecture_flutter/modules/home_module/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: 8b1a2f1016dcd9e5a71bbd4c743ecc9bb3c276a7
+  channel: stable
+
+project_type: plugin
