+++ software_architecture_flutter/modules/home_module/lib/home_module.dart
@@ -1,0 +1,4 @@
+library home_module;
+
+export 'src/home_module.dart';
+export 'src/home_page.dart';
