+++ software_architecture_flutter/modules/home_module/lib/src/components/custom_switch_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:software_architecture_flutter/app/controllers/app_controller.dart';
+import 'package:core_module/core_module.dart' show ChangeThemeViewmodel;
 
 // * Widget to request the AppController to modify the app's theme
 
@@ -10,12 +10,12 @@ class CustomSwitchWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ValueListenableBuilder(
-        valueListenable: Modular.get<AppController>().themeSwitch,
+        valueListenable: Modular.get<ChangeThemeViewmodel>().config.themeSwitch,
         builder: (context, isDark, child) {
           return Switch(
-            value: Modular.get<AppController>().isDark,
-            onChanged:
-                Modular.get<AppController>().changeThemeViewmodel.changeTheme,
+            value: Modular.get<ChangeThemeViewmodel>().config.themeSwitch ==
+                isDark,
+            onChanged: Modular.get<ChangeThemeViewmodel>().changeTheme,
             /*
             value: Modular.get<AppController>().isDark,
             onChanged: (bool value) {
