+++ software_architecture_flutter/modules/home_module/lib/src/controllers/home_controller.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
-import 'package:software_architecture_flutter/app/modules/home/models/apiAdvisor_model.dart';
-import 'package:software_architecture_flutter/app/modules/home/viewmodels/apiAdvisor_viewmodel.dart';
+import 'package:home_module/src/models/apiAdvisor_model.dart';
+import 'package:home_module/src/viewmodels/apiAdvisor_viewmodel.dart';
 
 class HomeController {
   final ApiAdvisorViewModel viewModel;
