+++ software_architecture_flutter/modules/home_module/lib/src/home_page.dart
@@ -1,15 +1,11 @@
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:software_architecture_flutter/app/modules/home/models/apiAdvisor_model.dart';
-import 'package:software_architecture_flutter/app/modules/home/controllers/home_controller.dart';
-import 'package:software_architecture_flutter/app/modules/home/repositories/apiAdvisor_repository.dart';
-import 'package:software_architecture_flutter/app/core/services/client_http_service.dart';
-import 'package:software_architecture_flutter/app/modules/home/viewmodels/apiAdvisor_viewmodel.dart';
+import 'package:home_module/src/components/custom_switch_widget.dart';
+import 'package:home_module/src/controllers/home_controller.dart';
+import 'package:home_module/src/models/apiAdvisor_model.dart';
 import 'package:test_api/test_api.dart';
 import 'package:http/http.dart';
-
 import 'package:flutter/material.dart';
-import 'package:software_architecture_flutter/app/modules/home/components/custom_switch_widget.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
