+++ software_architecture_flutter/modules/home_module/lib/src/repositories/apiAdvisor_repository.dart
@@ -1,6 +1,6 @@
-import 'package:software_architecture_flutter/app/core/interfaces/client_http_interface.dart';
-import 'package:software_architecture_flutter/app/modules/home/models/apiAdvisor_model.dart';
-import 'package:software_architecture_flutter/app/modules/home/repositories/apiAdvisor_repository_interface.dart';
+import 'package:core_module/core_module.dart';
+import 'package:home_module/src/models/apiAdvisor_model.dart';
+import 'package:home_module/src/repositories/apiAdvisor_repository_interface.dart';
 
 class ApiAdvisorRepository implements IApiAdvisor {
   final IClientHttp client;
