+++ software_architecture_flutter/modules/home_module/lib/src/repositories/apiAdvisor_repository_interface.dart
@@ -1,0 +1,5 @@
+import 'package:home_module/src/models/apiAdvisor_model.dart';
+
+abstract class IApiAdvisor {
+  Future<ApiAdvisorModel?> getWeather();
+}
